+++ src/main/java/racing/domain/FixedPowerGenerator.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 public class FixedPowerGenerator implements PowerGenerator {
     @Override
