+++ src/main/java/racing/domain/PowerGenerator.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 public interface PowerGenerator {
     int generate();
