+++ src/main/java/racing/domain/RacingGame.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 import racing.view.OutputView;
