+++ src/test/java/racing/domain/CarTest.java
@@ -1,8 +1,9 @@
-package racing;
+package racing.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import racing.domain.Car;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
